fix java skills to work around bug with extending functions with apply method in them.
