+++ function-aws-alexa/src/main/java/io/micronaut/function/aws/alexa/AlexaFunction.java
@@ -3,13 +3,14 @@ package io.micronaut.function.aws.alexa;
 import com.amazon.ask.Skill;
 import com.amazon.ask.model.RequestEnvelope;
 import com.amazon.ask.model.ResponseEnvelope;
+import io.micronaut.function.FunctionBean;
 
 import java.util.function.Function;
 
-public abstract class AlexFunction implements Function<RequestEnvelope, ResponseEnvelope> {
+public abstract class AlexaFunction implements Function<RequestEnvelope, ResponseEnvelope> {
 
     @Override
-    public final ResponseEnvelope apply(RequestEnvelope requestEnvelope) {
+    public ResponseEnvelope apply(RequestEnvelope requestEnvelope) {
         return getSkill().invoke(requestEnvelope);
     }
