+++ function-aws-alexa/src/test/groovy/io/micronaut/function/aws/alexa/handlers/HelpIntentHandler.java
@@ -11,7 +11,7 @@
      the specific language governing permissions and limitations under the License.
 */
 
-package hello.world.lambda.java.helloworld.handlers;
+package io.micronaut.function.aws.alexa.handlers;
 
 import com.amazon.ask.dispatcher.request.handler.HandlerInput;
 import com.amazon.ask.dispatcher.request.handler.RequestHandler;
@@ -21,18 +21,20 @@ import java.util.Optional;
 
 import static com.amazon.ask.request.Predicates.intentName;
 
-public class CancelandStopIntentHandler implements RequestHandler {
+public class HelpIntentHandler implements RequestHandler {
+
     @Override
     public boolean canHandle(HandlerInput input) {
-        return input.matches(intentName("AMAZON.StopIntent").or(intentName("AMAZON.CancelIntent")));
+        return input.matches(intentName("AMAZON.HelpIntent"));
     }
 
     @Override
     public Optional<Response> handle(HandlerInput input) {
-        String speechText = "Goodbye";
+        String speechText = "You can say hello to me!";
         return input.getResponseBuilder()
                 .withSpeech(speechText)
                 .withSimpleCard("HelloWorld", speechText)
+                .withReprompt(speechText)
                 .build();
     }
 }
