+++ function-groovy/src/test/groovy/io/micronaut/function/groovy/RoundFunction.groovy
@@ -25,3 +25,12 @@ math.multiplier = 2
 int round(float value) {
     mathService.round(value) // go
 }
+
+
+void kinesisEvent(KinesisEvent event) {
+
+
+
+
+
+}
